+++ sei-ibc-go/modules/light-clients/07-tendermint/types/genesis.go
@@ -6,11 +6,11 @@ import (
 	"github.com/cosmos/ibc-go/modules/core/exported"
 )
 
-// ExportMetadata exports all the processed times in the client store so they can be included in clients genesis
+// ExportMetadata exports all the consensus metadata in the client store so they can be included in clients genesis
 // and imported by a ClientKeeper
 func (cs ClientState) ExportMetadata(store sdk.KVStore) []exported.GenesisMetadata {
 	gm := make([]exported.GenesisMetadata, 0)
-	IterateProcessedTime(store, func(key, val []byte) bool {
+	IterateConsensusMetadata(store, func(key, val []byte) bool {
 		gm = append(gm, clienttypes.NewGenesisMetadata(key, val))
 		return false
 	})